alarmtimer: add support to set the maximum alarm interval time
When system enters into suspend, alarmtimer set the RTC HW for alarm.
Add API to set the maximum time for alarm such that if any client/user
space has set the wakeup for more than this time then alarmtimer will
set the alarm to this time.

This help to ensure that system will wakeup once on this time.

bug 200082198

Change-Id: Ib22ad0a01782e946750a6641579bcad9f119fd70
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: http://git-master/r/720942
Reviewed-on: http://git-master/r/722702
Tested-by: Saket Bhojane <sbhojane@nvidia.com>